opt_bagging/requirements.txt
tensorflow
xmanager
numpy
scikit-learn
scipy
absl-py
tensorflow-probability
